+++ src/events/index.ts
@@ -1,5 +1,0 @@
-import EventEmitter from "./Emitter";
-
-const eventEmitter = new EventEmitter();
-
-export default eventEmitter;