+++ src/index.ts
@@ -1,3 +1,7 @@
+import { Layer } from "./routing";
+
+export type Handler = Layer.Handler;
+
 export * from "./constants";
 export { HttpException } from "./exceptions";
 export { Router } from "./routing";